refactor: 객체 할당 시점 초기화 로직 적용
SparsePool에서 객체 인스턴스화 시 내부 데이터 리셋 실행
이전 세션의 잔류 데이터로 인한 논리적 버그 및 메모리 오염 차단

+++ Network/Common/SparseSet.hpp
@@ -12,7 +12,7 @@ template <size_t N>
 class SparseSet {
    public:
 	struct Slot {
-		uint32_t where_ = 0;
+		size_t where_ = 0;
 		uint32_t generation_ = 0;
 	};
